This change adds an exclude option to the middleman-cdn plugin gem.  This option permits an array of regular expressions to be provided which will not be invalidated.
The chief purpose of this is to permit interoperability with deploying your build
directory with git (such as middleman-deploy does), an exclude operation is provided
to allow a reduction in scope of the files specified by filter.  This configuration
option is an Array of regular expressions (Regex) objects.  When expanding the files
to invalid, if any of the filenames (without leading /) match any of the regular
expressions, then that file is not validated.

    activate :cdn do |cdn|
      cdn.cloudflare = {
        client_api_key: '...',   # default ENV['CLOUDFLARE_CLIENT_API_KEY']
        email: '...',            # default ENV['CLOUDFLARE_EMAIL']
        zone: '...',
        base_urls: [ ... ]
      }

      # This excludes the entire .git repo AND and .keep files in empty directories.
      cdn.exclude = [/^.git/, /.keep$/]
    end

Also included herein are the corresponding changes to documentation and a test.